requirements.txt
streamlit
opencv-python
numpy
Pillow
sentence-transformers
psycopg2-binary
streamlit-lottie
openai
